refactored phrase with tool tip
